refactor(python): mypy 全量类型修复与注解补全
- app 核心目录 mypy 错误 1467→51（ai/budget/dreaming/rag/database 等 23 目录清零）
- 修复真实 bug：RepositoryError.ValidationError 参数 TypeError、cutting_params_db 未知材料 KeyError
- 修复 sandbox worker 缺 binascii import、startup_hooks alembic 迁移改 asyncio.to_thread
- 补充嵌套字面量类型别名（OperationParams/MaterialOps/MaterialParams）等 ~200 处注解

+++ engineering/python/app/ai/unified_embedding/encoder.py
@@ -317,7 +317,7 @@ class MultiModalEncoder:
         for emb, w in zip(embeddings, normalized_weights):
             if emb.ndim == 2:
                 emb = emb[0]
-            fused = fused + emb * w
+            fused = fused + (emb * w).astype(np.float32)
         return self._space.normalize(fused)
 
     def fuse_batch(
@@ -350,7 +350,7 @@ class MultiModalEncoder:
 
         fused = np.zeros((batch_size, TOTAL_DIMS), dtype=np.float32)
         for emb, w in zip(active, normalized_weights):
-            fused = fused + emb * w
+            fused = fused + (emb * w).astype(np.float32)
         return self._space.normalize(fused)
 
     def encode_manufacturing_scene(
